feat: adjustment in create table orders

+++ src/main/resources/db/migration/V2__Create_Table_Orders.sql
@@ -1,1 +1,13 @@
-create table orders (order_id bigint not null auto_increment, conclusion_date datetime(6), created_at datetime(6), is_done bit, is_paid bit, order_code binary(16), order_details varchar(255), order_total float(53), status smallint, primary key (order_id));
+create table orders
+(
+    order_id        bigint not null auto_increment,
+    conclusion_date datetime(6),
+    created_at      datetime(6),
+    is_done         bit,
+    is_paid         bit,
+    order_code      binary(16),
+    order_details   varchar(255),
+    order_total     float(53),
+    status          int,
+    primary key (order_id)
+);
